feat: remove old grid code
