+++ packages/graphql-toolbox/src/modules/EditorView/grid/grid.css
@@ -1,88 +1,23 @@
-.react-resizable {
-    position: relative;
+.horizontal-drag-bar {
+  cursor: col-resize;
+  bottom: 0;
+  right: 0;
+  width: 15px;
+  background-color: transparent;
+  background-repeat: no-repeat;
+  background-origin: content-box;
+  box-sizing: border-box;
+  background-position: center;
 }
 
-.react-resizable-handle {
-    position: absolute;
-    width: 100%;
-    height: 100%;
-}
-  
-.react-resizable-handle-w {
-    left: 0;
-    top: 0;
-    width: 15px;
-    cursor: ew-resize;
-    background-color: transparent;
-    background-repeat: no-repeat;
-    background-origin: content-box;
-    box-sizing: border-box;
-    background-position: center;
-}
- 
-.react-resizable-handle-s {
-    bottom: 0;
-    right: 0;
-    height: 15px;
-    cursor: ns-resize;
-    background-color: transparent;
-    background-repeat: no-repeat;
-    background-origin: content-box;
-    box-sizing: border-box;
-    background-position: center;
-}
-
-.react-resizable-handle-se {
-    bottom: 0;
-    right: 0;
-    height: 0;
-}
-
-.the-grid {
-    position: relative;
-    display: grid;
-    grid-template-columns: 1fr 0.3fr;
-    grid-template-rows: 1fr 1fr;
-    grid-gap: 0px;
-
-    width: 100%;
-    height: 90%;
-}
-
-.right {
-    grid-column: 2;
-    grid-row: 1 / span 2;
-}
-
-.left-top {
-    grid-column: 1;
-    grid-row: 1;
-}
-
-.left-bottom {
-    grid-column: 1;
-    grid-row: 2;
-}
-
-.right-inner {
-    padding-left: 15px;
-    width: 100%;
-    min-width: 100%;
-    height: 100%;
-    min-height: 100%;
-}
-
-.left-top-inner {
-    padding-bottom: 65px;
-    width: 100%;
-    min-width: 100%;
-    height: 100%;
-    min-height: 100%;
-}
-
-.left-bottom-inner {
-    width: 100%;
-    min-width: 100%;
-    height: 100%;
-    min-height: 100%;
+.vertical-drag-bar {
+  cursor: row-resize;
+  left: 0;
+  top: 0;
+  height: 15px;
+  background-color: transparent;
+  background-repeat: no-repeat;
+  background-origin: content-box;
+  box-sizing: border-box;
+  background-position: center;
 }
