--- packages/graphql-toolbox/src/modules/EditorView/twogrid/Grid.css
@@ -1,23 +1,0 @@
-.horizontal-drag-bar {
-  cursor: col-resize;
-  bottom: 0;
-  right: 0;
-  width: 15px;
-  background-color: transparent;
-  background-repeat: no-repeat;
-  background-origin: content-box;
-  box-sizing: border-box;
-  background-position: center;
-}
-
-.vertical-drag-bar {
-  cursor: row-resize;
-  left: 0;
-  top: 0;
-  height: 15px;
-  background-color: transparent;
-  background-repeat: no-repeat;
-  background-origin: content-box;
-  box-sizing: border-box;
-  background-position: center;
-}
